backend/pyproject.toml
[project]
name = "AutoBangumi"
version = "3.1.14"
description = "AutoBangumi 是基于 RSS 的全自动追番整理下载工具。只需要在 Mikan Project 等网站上订阅番剧，就可以全自动追番。 并且整理完成的名称和目录可以直接被 Plex、Jellyfin 等媒体库软件识别，无需二次刮削。"
dependencies = [
    "sqlmodel>=0.0.21",
    "torrentool>=1.2.0",
    "pydantic>=2.8.2",
    "httpx[socks,http2]>=0.27.0",
    "lxml>=5.2.2",
    "openai>=1.37.0",
    "python-dotenv>=1.0.1",
    "python-jose>=3.3.0",
    "qbittorrent-api>=2024.7.64",
    "bs4>=0.0.2",
    "semver>=3.0.2",
    "sse-starlette>=2.1.2",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt==4.0.1",
    "fastapi-slim>=0.111.1",
    "uvicorn>=0.30.3",
    "python-multipart>=0.0.9",
    "jinja2>=3.1.4",
    "loguru>=0.7.2",
]
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "ruff>=0.5.4",
    "black>=24.4.2",
    "pre-commit>=3.7.1",
    "pytest>=8.3.1",
]
virtual = true

[tool.rye.scripts]
dev = "bash dev.sh"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/module"]

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]


# Same as Black.
line-length = 88


# Assume Python 3.10.
target-version = "py310"

[tool.ruff.lint]
# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"
# Allow autofix for all enabled rules (when `--fix`) is provided.
fixable = [
    "A",
    "B",
    "C",
    "D",
    "E",
    "F",
    "G",
    "I",
    "N",
    "Q",
    "S",
    "T",
    "W",
    "ANN",
    "ARG",
    "BLE",
    "COM",
    "DJ",
    "DTZ",
    "EM",
    "ERA",
    "EXE",
    "FBT",
    "ICN",
    "INP",
    "ISC",
    "NPY",
    "PD",
    "PGH",
    "PIE",
    "PL",
    "PT",
    "PTH",
    "PYI",
    "RET",
    "RSE",
    "RUF",
    "SIM",
    "SLF",
    "TCH",
    "TID",
    "TRY",
    "UP",
    "YTT",
]
unfixable = []
ignore = [
    # E501: https://beta.ruff.rs/docs/rules/line-too-long/
    'E501',
    # F401: https://beta.ruff.rs/docs/rules/unused-import/
    # avoid unused imports lint in `__init__.py`
    'F401',
]

select = [
    # pycodestyle(E): https://beta.ruff.rs/docs/rules/#pycodestyle-e-w
    "E",
    # Pyflakes(F): https://beta.ruff.rs/docs/rules/#pyflakes-f
    "F",
    # isort(I): https://beta.ruff.rs/docs/rules/#isort-i
    "I",
]
per-file-ignores = {}

[tool.ruff.lint.mccabe]
# Unlike Flake8, default to a complexity level of 10.
max-complexity = 10

[tool.black]
line-length = 88
target-version = ['py310', 'py311']
